// src/typings/InvestingNews.ts
export interface InvestingNews {
  article_ID: number;
  title: string;
  shortTitle: string;
  href: string;
  imageHref: string;
  provider: string;
  providerHref: string;
  date: string;
  commentsCounter: number;
  snippet: string;
  openInNewTab: boolean;
  mediumImageHref: string;
  smallImageHref: string;
  news_type: string;
  category: string;
  categoryLink: string;
  article_type: string;
}